refactor: update SymbolTable
