+++ packages/shader-lab/src/common/IBaseSymbol.ts
@@ -1,0 +1,3 @@
+export interface IBaseSymbol {
+  readonly ident: string;
+}
